Finalize storage
